api/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
openpyxl
python-pptx
PyMuPDF
google-generativeai
python-dotenv
